Work around incomplete iterators >:(
Anyone who uses json.hpp and src/json.o would not be able to instantiate
the incomplete iterators (see https://stackoverflow.com/a/57930951).
Therefore, because the header cannot be modified (as for specification),
I had to include src/json.cpp directly (thus breaking the pimpl idiom).

If the header could be changed, the simplest solution I found would be
to completely avoid the implementation and use public methods to iterate
the json (see https://softwareengineering.stackexchange.com/a/264916).